+++ client/src/main/java/org/eclipse/hono/client/HonoClient.java
@@ -255,6 +255,19 @@ public interface HonoClient {
     Future<TenantClient> getOrCreateTenantClient();
 
     /**
+     * Gets a client for sending command with Hono's <em>Command and Control</em> API.
+     *
+     * @param tenantId The tenant to the all command receiving devices belong.
+     * @param deviceId The device id.
+     * @return A future that will complete with the command and control client (if successful) or
+     *         fail if the client cannot be created, e.g. because the underlying connection
+     *         is not established or if a concurrent request to create a client for the same
+     *         tenant and device is already being executed.
+     * @throws NullPointerException if the tenant is {@code null}.
+     */
+    Future<CommandClient> getOrCreateCommandClient(String tenantId, String deviceId);
+
+    /**
      * Closes this client's connection to the Hono server.
      * <p>
      * This method waits for at most 5 seconds for the connection to be closed properly. Any subsequent attempts to